Move all function/methods to CPP files
 1) Moved all methods to the implementation files, headers are quite
    empty now.
 2) Kept simple methods inline, because "this is the way"
 3) Tried to use `const` where possible.

Did not test the whole app - things might break, this needs a few
moments of testing.
