+++ AppSettings.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "config.h"
+
 #include <QApplication>
 #include <QDir>
 #include <QFont>
@@ -22,168 +22,25 @@ class AppSettings : public QObject
     Q_PROPERTY(bool useClipboard READ useClipboard WRITE setUseClipboard NOTIFY useClipboardChanged)
     Q_PROPERTY(int fontSize READ fontSize WRITE setfontSize NOTIFY fontSizeChanged)
 
-    // hygen Q_PROPERTY
     QML_ELEMENT
 public:
-    AppSettings(QObject *parent = nullptr)
-        : QObject(parent)
-    {
-        m_passwordStorePath = settings.value("passwordStorePath", "").toString();
-        m_tmpFolderPath = settings.value("tmpFolderPath", "").toString();
-        m_gitExecPath = settings.value("gitExecPath", "").toString();
-        m_vscodeExecPath = settings.value("vscodeExecPath", "").toString();
-        m_autoTypeCmd = settings.value("autoTypeCmd", "").toString();
-        m_ctxSigner = settings.value("ctxSigner", "").toString();
-        m_useClipboard = settings.value("useClipboard", "").toBool();
-        m_fontSize = settings.value("fontSize", "14").toInt();
-    }
-
-    QString passwordStorePath()
-    {
-        QString passwordStorePathDefault = QDir::homePath() + "/.password-store";
-        if (m_passwordStorePath.isEmpty() || !QDir(m_passwordStorePath).exists()) {
-            return passwordStorePathDefault;
-        }
-
-        return m_passwordStorePath;
-    };
-
-    void setPasswordStorePath(const QString &passwordStorePath)
-    {
-        if (passwordStorePath == m_passwordStorePath)
-            return;
-
-        m_passwordStorePath = passwordStorePath;
-        settings.setValue("passwordStorePath", m_passwordStorePath);
-    }
-
-    QString tmpFolderPath()
-    {
-        QString tmpFolderPathDefault = QDir::tempPath();
-
-        if (QString(PROJECT_OS) == "LINUX" && QDir("/dev/shm").exists()) {
-            tmpFolderPathDefault = "/dev/shm";
-        }
-
-        if (m_tmpFolderPath.isEmpty() || !QDir(m_tmpFolderPath).exists()) {
-            return tmpFolderPathDefault;
-        }
-        return m_tmpFolderPath;
-    };
-
-    void setTmpFolderPath(const QString &tmpFolderPath)
-    {
-        if (tmpFolderPath == m_tmpFolderPath)
-            return;
-
-        m_tmpFolderPath = tmpFolderPath;
-        settings.setValue("tmpFolderPath", m_tmpFolderPath);
-    }
-
-    QString gitExecPath()
-    {
-        if (m_gitExecPath.isEmpty()) {
-            return "git";
-        }
-        return m_gitExecPath;
-    };
-
-    void setGitExecPath(const QString &gitExecPath)
-    {
-        if (gitExecPath == m_gitExecPath)
-            return;
-
-        m_gitExecPath = gitExecPath;
-        settings.setValue("gitExecPath", m_gitExecPath);
-
-        emit gitExecPathChanged();
-    }
-
-    QString ctxSigner() { return m_ctxSigner; };
-
-    void setCtxSigner(const QString &ctxSigner)
-    {
-        if (ctxSigner == m_ctxSigner)
-            return;
-
-        m_ctxSigner = ctxSigner;
-        settings.setValue("ctxSigner", m_ctxSigner);
-
-        emit ctxSignerChanged();
-    }
-
-    QString vscodeExecPath()
-    {
-        if (m_vscodeExecPath.isEmpty()) {
-            if (QString(PROJECT_OS) == "LINUX") {
-                return "/usr/bin/code";
-            }
-            return "/usr/local/bin/code";
-        }
-        return m_vscodeExecPath;
-    };
-
-    void setVscodeExecPath(const QString &vscodeExecPath)
-    {
-        if (vscodeExecPath == m_vscodeExecPath)
-            return;
-
-        m_vscodeExecPath = vscodeExecPath;
-        settings.setValue("vscodeExecPath", m_vscodeExecPath);
-
-        emit vscodeExecPathChanged();
-    }
-
-    QString autoTypeCmd()
-    {
-        if (m_autoTypeCmd.isEmpty()) {
-            if (QString(PROJECT_OS) == "LINUX") {
-                return R"V0G0N(
-echo -n sequence | xclip -selection clipboard
-    )V0G0N";
-            }
-            return "";
-        }
-        return m_autoTypeCmd;
-    };
-
-    void setAutoTypeCmd(const QString &autoTypeCmd)
-    {
-        if (autoTypeCmd == m_autoTypeCmd)
-            return;
-
-        m_autoTypeCmd = autoTypeCmd;
-        settings.setValue("autoTypeCmd", m_autoTypeCmd);
-
-        emit autoTypeCmdChanged();
-    }
-
-    bool useClipboard() { return m_useClipboard; };
-
-    void setUseClipboard(const bool useClipboard)
-    {
-        if (useClipboard == m_useClipboard)
-            return;
-
-        m_useClipboard = useClipboard;
-        settings.setValue("useClipboard", m_useClipboard);
-
-        emit useClipboardChanged();
-    }
-
-    int fontSize() { return m_fontSize; };
-
-    void setfontSize(const int fontSize)
-    {
-        if (fontSize == m_fontSize)
-            return;
-
-        m_fontSize = fontSize;
-        settings.setValue("fontSize", m_fontSize);
-
-        emit fontSizeChanged();
-    }
-    // hygen public
+    AppSettings(QObject *parent = nullptr);
+    const QString passwordStorePath() const;
+    void setPasswordStorePath(const QString &passwordStorePath);
+    const QString tmpFolderPath() const;
+    void setTmpFolderPath(const QString &tmpFolderPath);
+    const QString gitExecPath() const;
+    void setGitExecPath(const QString &gitExecPath);
+    const QString ctxSigner() const { return m_ctxSigner; };
+    void setCtxSigner(const QString &ctxSigner);
+    QString vscodeExecPath() const;
+    void setVscodeExecPath(const QString &vscodeExecPath);
+    const QString autoTypeCmd() const;
+    void setAutoTypeCmd(const QString &autoTypeCmd);
+    bool useClipboard() const { return m_useClipboard; };
+    void setUseClipboard(const bool useClipboard);
+    int fontSize() const { return m_fontSize; };
+    void setfontSize(const int fontSize);
 
 signals:
     void passwordStorePathChanged();
@@ -195,8 +52,6 @@ signals:
     void fontSizeChanged();
     void ctxSignerChanged();
 
-    // hygen signals
-
 private:
     QSettings settings{"shemeshg", "PassSimple"};
     QString m_passwordStorePath;
@@ -207,6 +62,4 @@ private:
     QString m_ctxSigner;
     bool m_useClipboard;
     int m_fontSize;
-
-    // hygen private
 };
