+++ InputType/InputTypeType.h
@@ -10,18 +10,13 @@
 class InputTypeType : public QObject
 {
     Q_OBJECT
-    // hygen Q_PROPERTY
     QML_ELEMENT
 
 public:
     explicit InputTypeType(QObject *parent = nullptr)
         : QObject(parent){};
 
-    // hygen public
-
 signals:
-    // hygen signals
 
 private:
-    // hygen private
 };
