+++ mainwindow.h
@@ -1,7 +1,5 @@
-#ifndef MAINWINDOW_H
-#define MAINWINDOW_H
+#pragma once
 
-#include "UiGuard.h"
 #include "config.h"
 #include "mainqmltype.h"
 #include <QFileIconProvider>
@@ -26,26 +24,12 @@ public:
     ~MainWindow();
 
 public slots:
-    void indexHasChanged()
-    {
-        UiGuard guard(this);
-        // get the text of the selected item
-
-        auto idx = treeIndex.model()->index(treeIndex.row(), 0, treeIndex.parent());
-        try {
-            mainqmltype->setFilePath(filesystemModel.filePath(idx));
-        } catch (const std::exception &e) {
-            qDebug() << e.what();
-        }
-    }
+    void indexHasChanged();
 
 private slots:
-
     void on_splitter_splitterMoved(int pos, int index);
-
     void selectionChangedSlot(const QItemSelection & /*newSelection*/,
                               const QItemSelection & /*oldSelection*/);
-
     void about() { QMessageBox::about(this, tr("About Menu"), tr("Version %1").arg(PROJECT_VER)); }
 
 private:
@@ -55,4 +39,3 @@ private:
     MainQmlType *mainqmltype;
     QModelIndex treeIndex;
 };
-#endif // MAINWINDOW_H
